Rename app from gameini to game_app
