+++ gameini/game_app/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.conf.urls import url
 from django.contrib import admin
-from gameini.views import placeholder
+from game_app.views import placeholder
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
