+++ gameini/game_app/wsgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "gameini.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "game_app.settings")
 
 application = get_wsgi_application()
